+++ lib/yard.rb
@@ -48,6 +48,9 @@ module YARD
 
   # @return [Boolean] whether YARD is being run in Ruby 2.0
   def self.ruby2?; @ruby2 ||= (RUBY_VERSION >= '2.0.0') end
+
+  # @return [Boolean] whether YARD is being run in Ruby 3.0
+  def self.ruby3?; @ruby3 ||= (RUBY_VERSION >= '3.0.0') end
 end
 
 # Keep track of Ruby version for compatibility code