+++ lib/yard/parser/ruby/ast_node.rb
@@ -480,7 +480,7 @@ module YARD
         end
 
         def parameters(include_block_param = true)
-          params = self[1 + index_adjust]
+          return unless params = self[1 + index_adjust]
           params = params[0] if params.type == :paren
           include_block_param ? params : params[0...-1]
         end
@@ -488,7 +488,7 @@ module YARD
         def signature
           params_src = ''
           params = self[1 + index_adjust]
-          if params.first
+          if params and params.first
             params_src = params.type == :paren ? '' : ' '
             params_src += params.source.gsub(/\s+(\s|\))/m, '\1')
           end