+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # main
 
+- Add support for Ruby 3.0 endless method definitions. (#1376)
+
 # 0.9.26 - December 26th, 2020
 
 [0.9.26]: https://github.com/lsegal/yard/compare/v0.9.25...v0.9.26